Move host bindings to compiler
The compiler is responsible for adding types and checking for undefined
variables, that sort of thing. Or it would be if I bothered to implement
it. But when I get there it will need to know what bindings are
available so it doesn't flag them as undefined.

This also means I don't have to use the `interpreter` function directly
and I can pass everything through `run`, which might be the real reason
I did this.